+++ lib/provider.dart
@@ -1,5 +1,7 @@
 library provider;
 
+export 'src/adaptive_builder_widget.dart' show ValueBuilder;
 export 'src/consumer.dart';
 export 'src/listenable_provider.dart';
-export 'src/provider.dart' hide debugGetProviderUpdateShouldNotify;
+export 'src/provider.dart';
+export 'src/stream_provider.dart';