+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 2.0.0
 
+- merged `Provider` and `StatefulProvider`
+- added a "builder" constructor to `ValueListenableProvider`
+- normalized providers constructors such that the default constructor is a "builder", and offer a `value` named constructor.
+
 # 1.6.1
 
 - `Provider.of<T>` now crashes with a `ProviderNotFoundException` when no `Provider<T>` are found in the ancestors of the context used.